requirements.txt
python
psycopg2-binary
fastapi